feat(ui): migrate search feature module
