+++ apps/ui/app/product/ProductPageClient.tsx
@@ -15,8 +15,8 @@ import { useCategories } from '@/lib/hooks/useCategories';
 import { mapApiProductToUiProduct } from '@/lib/utils/productMappers';
 import { formatAgentResponse, type AgentMessageView } from '@/lib/utils/agentResponseCards';
 import AgentMessageDisplay from '@/components/organisms/AgentMessageDisplay';
-import { UseCaseTags } from '@/components/enrichment/UseCaseTags';
-import { RelatedProductsRail } from '@/components/enrichment/RelatedProductsRail';
+import { RelatedProductsRail } from '@/src/shared/components/RelatedProductsRail';
+import { UseCaseTags } from '@/src/shared/components/UseCaseTags';
 import { useRelatedProducts } from '@/lib/hooks/useRelatedProducts';
 import agentApiClient from '@/lib/api/agentClient';
 import { recordAgentInvocationTelemetry } from '@/lib/hooks/useAgentInvocationTelemetry';
