+++ apps/ui/components/demo/ExecutiveDemoPage.tsx
@@ -27,8 +27,8 @@ import { useOrders } from '@/lib/hooks/useOrders';
 import { useProductSimilarity } from '@/lib/hooks/useProductSimilarity';
 import { useProducts } from '@/lib/hooks/useProducts';
 import { useReturns } from '@/lib/hooks/useReturns';
-import { useStreamingSearch } from '@/lib/hooks/useStreamingSearch';
 import { useTruthAnalyticsSummary } from '@/src/features/truth';
+import { useStreamingSearch } from '@/src/features/search';
 import type { AgentHealthCardMetric, AgentModelUsageRow } from '@/lib/types/api';
 import { formatAgentResponse } from '@/lib/utils/agentResponseCards';
 import { mapApiProductsToUi } from '@/lib/utils/productMappers';
