+++ apps/ui/src/features/search/components/SearchComparisonScorecard.tsx
