+++ apps/ui/src/features/search/hooks/useStreamingSearch.ts
