+++ apps/ui/src/features/search/internal/appPages.ts
