+++ apps/ui/src/features/search/types/index.ts
@@ -1,0 +1,25 @@
+export type {
+  SearchDegradedReason,
+  SearchModePreference,
+  SearchResultType,
+  SemanticSearchContext,
+  SemanticSearchIntent,
+  SemanticSearchRequest,
+  SemanticSearchResponse,
+  StreamingSearchCallbacks,
+} from '../services/semanticSearchService';
+
+export type {
+  IntelligentSearchOptions,
+  IntelligentSearchPreference,
+  IntelligentSearchStage,
+} from '../hooks/useIntelligentSearch';
+
+export type { AppAudience, AppPage } from '../internal/appPages';
+export type { SearchHit } from '../internal/matcher';
+export type { AppSearchBoxProps } from '../components/AppSearchBox';
+export type { SearchComparisonItem } from '../components/SearchComparisonScorecard';
+export type { SearchResultCardProps } from '../components/SearchResultCard';
+export type { SearchModeIndicatorProps } from '../components/SearchModeIndicator';
+export type { SearchModeToggleProps } from '../components/SearchModeToggle';
+export type { IntentPanelProps } from '../components/IntentPanel';
