+++ apps/ui/src/shared/components/ConfidenceBadge.tsx
