+++ apps/ui/src/shared/components/RelatedProductsRail.tsx
@@ -1,7 +1,7 @@
 import React from 'react';
 import Link from 'next/link';
 import Image from 'next/image';
-import type { Product } from '../types';
+import type { Product } from '@/components/types';
 
 export interface RelatedProductsRailProps {
   title: string;
